Make print for for python2 and python3
As far as I can see, the only thing in the script that does not work in python3 are the print statements. If we enclose all the strings in parenthesis, they will be treated as a function call in python3 and will work just fine in python2 because they are treated as an expression. Tested with 2.7.9 and 3.3.5. This would make the scipt usable in systems that only have python3. This is an alternative to #5