Add travis CI integration
While the current version of the script works with both python2 and
python3, we should ensure that future changes do not break for either
python2 or python3.

The test only calls oclint-json-compilation-database -v for now. In the
future we might want to test against the actual oclint binary using real
compilation database files.